requirements.txt
async-timeout==4.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
importlib-metadata==6.7.0
python-dotenv==0.21.1
redis==5.0.1
requests==2.31.0
typing-extensions==4.7.1
urllib3==2.0.7
zipp==3.15.0
